JDG-1892 ISPN-9378 Avoid unneeded parentheses in Ickle query string generated by QueryBuilder
Each QueryStringCreator.visit should generate the predicate and wrap it
with ( ) depending on the type of the parent, a1892voiding the wrapping if
it makes sense. The responsibility for wrapping was previosly in the
parent, but moving it to the child can generate better code, with no unwanted
parentheses while still maintaining operator precedence.